Move query object contracts to queries namespace
also define contract on each query object instead of guess it in the
base query class.
