--- app/contracts/billable_metrics_query_filters_contract.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-class BillableMetricsQueryFiltersContract < Dry::Validation::Contract
-  params do
-    required(:filters).hash do
-      optional(:recurring).filled(:bool)
-      optional(:aggregation_types).array(:string, included_in?: %w[max_agg count_agg])
-    end
-
-    optional(:search_term).maybe(:string)
-  end
-end
